Exclude spatial from 2.x tests
