+++ core/src/test/java/org/neo4j/ogm/domain/cineasts/annotated/User.java
@@ -16,7 +16,8 @@ package org.neo4j.ogm.domain.cineasts.annotated;
 
 import org.neo4j.ogm.annotation.Relationship;
 import org.neo4j.ogm.annotation.typeconversion.Convert;
-import org.neo4j.ogm.domain.cineasts.utils.LocationConverter;
+import org.neo4j.ogm.domain.restaurant.LocationConverter;
+import org.neo4j.ogm.domain.restaurant.Location;
 
 import java.net.URL;
 import java.util.List;
@@ -47,9 +48,6 @@ public class User {
     URL[] urls;
 
     String[] nicknames;
-
-    @Convert(LocationConverter.class)
-    Location location;
 
     Rating rate(Movie movie, int stars, String comment) {
         return null;
@@ -136,14 +134,6 @@ public class User {
 
     public void setNicknames(String[] nicknames) {
         this.nicknames = nicknames;
-    }
-
-    public Location getLocation() {
-        return location;
-    }
-
-    public void setLocation(Location location) {
-        this.location = location;
     }
 
     @Override
