+++ pom.xml
@@ -252,6 +252,36 @@
             <properties>
                 <neo4j>2.2.10</neo4j>
             </properties>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <version>2.18.1</version>
+                        <executions>
+                            <execution>
+                                <id>default-test</id>
+                                <configuration>
+                                    <forkMode>once</forkMode>
+                                    <includes>
+                                        <include>**/*.java</include>
+                                    </includes>
+                                    <excludes>
+                                        <exclude>**/defects/**</exclude>
+                                        <exclude>**/spatial/**</exclude>
+                                        <exclude>**/BoltDriverTest.java</exclude>
+                                    </excludes>
+                                    <systemPropertyVariables>
+                                        <ogm.properties>${ogm.properties}</ogm.properties>
+                                    </systemPropertyVariables>
+
+                                </configuration>
+                            </execution>
+                        </executions>
+                    </plugin>
+                </plugins>
+            </build>
+
         </profile>
 
         <profile>
@@ -259,6 +289,36 @@
             <properties>
                 <neo4j>2.3.6</neo4j>
             </properties>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <version>2.18.1</version>
+                        <executions>
+                            <execution>
+                                <id>default-test</id>
+                                <configuration>
+                                    <forkMode>once</forkMode>
+                                    <includes>
+                                        <include>**/*.java</include>
+                                    </includes>
+                                    <excludes>
+                                        <exclude>**/defects/**</exclude>
+                                        <exclude>**/spatial/**</exclude>
+                                        <exclude>**/BoltDriverTest.java</exclude>
+                                    </excludes>
+                                    <systemPropertyVariables>
+                                        <ogm.properties>${ogm.properties}</ogm.properties>
+                                    </systemPropertyVariables>
+
+                                </configuration>
+                            </execution>
+                        </executions>
+                    </plugin>
+                </plugins>
+            </build>
+
         </profile>
 
         <profile>
